+++ test/lsp4clj/server_test.clj
@@ -133,7 +133,7 @@
     (is (nil? (server/send-notification server "req" {:body "foo"})))
     (server/shutdown server)))
 
-(deftest should-receive-receive-response-to-request-sent-while-processing-notification
+(deftest should-receive-response-to-request-sent-while-processing-notification
   ;; https://github.com/clojure-lsp/clojure-lsp/issues/1500
   (let [input-ch (async/chan 3)
         output-ch (async/chan 3)
@@ -163,6 +163,57 @@
         (async/put! input-ch (lsp.responses/response (:id client-rcvd-request) {:processed true})))
       (is (= {:processed true} (deref client-resp 10000 :timed-out))))
     (server/shutdown server)))
+
+(deftest should-fail-pending-requests-if-too-many-inbound-messages-are-buffered
+  ;; * If the server sends a request, and blocks waiting for the response,
+  ;; * and if the client sends too many other messages before responding,
+  ;; * then the server's buffer of unprocessed inbound messages will fill up.
+  ;; To avoid dropping messages or buffering endlessly, the server eventually
+  ;; aborts its request.
+  (with-redefs [server/input-buffer-size 5]
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          client-req-id* (atom 0)
+          client-req (fn [body]
+                       (async/put! input-ch (lsp.requests/request (swap! client-req-id* inc)
+                                                                  "client-sent-request"
+                                                                  body)))]
+      (server/start server nil)
+      (with-redefs [server/receive-request (fn [_ _ {:keys [server-action] :as client-req}]
+                                             (if (= :block server-action)
+                                               (let [req (server/send-request server "server-sent-request" {:body "foo"})]
+                                                 (try
+                                                   (deref req)
+                                                   {:processed client-req}
+                                                   (catch Throwable _
+                                                     {:error {:result :deref-aborted}})))
+                                               {:processed client-req}))]
+        ;; The client sends a request which causes the server to send its own
+        ;; request. The server starts blocking, waiting for the client to respond.
+        (client-req {:server-action :block})
+        ;; The client receives the server's request but doesn't respond yet.
+        (is (= "server-sent-request" (:method (h/assert-take output-ch))))
+        ;; Before responding to the server's request, the client sends many other
+        ;; messages. The server will buffer these messages.
+        (dotimes [n server/input-buffer-size]
+          (client-req {:server-action :buffer, :input n}))
+        ;; The server is still blocking.
+        (is (h/assert-no-take output-ch))
+        ;; The client sends one more mesage, which is too many for the server to buffer.
+        (client-req {:server-action :overflow})
+        ;; To avoid blocking the client's inbound messages, the server's outbound
+        ;; request is aborted, causing it to stop waiting for a client response.
+        (is (= {:jsonrpc "2.0", :id 1, :error {:result :deref-aborted}}
+               (h/assert-take output-ch)))
+        ;; Now the server can process every other message from the client.
+        (dotimes [n server/input-buffer-size]
+          (is (= {:processed {:server-action :buffer, :input n}}
+                 (:result (h/assert-take output-ch)))))
+        (is (= {:processed {:server-action :overflow}}
+               (:result (h/assert-take output-ch)))))
+      (server/shutdown server))))
 
 (deftest should-cancel-request-when-cancellation-notification-receieved
   (let [input-ch (async/chan 3)
@@ -239,7 +290,7 @@
     (h/assert-take output-ch)
     (is (future-cancel req))
     (is (= "$/cancelRequest" (:method (h/assert-take output-ch))))
-    (is (not (future-cancel req)))
+    (is (future-cancel req))
     (h/assert-no-take output-ch)
     (server/shutdown server)))
 
@@ -314,6 +365,120 @@
       (future-cancel req)
       (is (thrown? java.util.concurrent.CancellationException
                    (.get req 100 java.util.concurrent.TimeUnit/MILLISECONDS)))
+      (server/shutdown server))))
+
+(deftest request-should-behave-like-a-promesa-promise
+  (testing "before being handled"
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          _ (server/start server nil)
+          req (p/promise (server/send-request server "req" {:body "foo"}))]
+      (is (not (p/done? req)))
+      (server/shutdown server)))
+  (testing "after response"
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          _ (server/start server nil)
+          req (p/promise (server/send-request server "req" {:body "foo"}))
+          client-rcvd-msg (h/assert-take output-ch)]
+      (async/put! input-ch (lsp.responses/response (:id client-rcvd-msg) {:result "good"}))
+      (is (= {:result :client-success
+              :value 2
+              :resp {:result "good"}}
+             (-> req
+                 (p/then (fn [resp] {:result :client-success
+                                     :value 1
+                                     :resp resp}))
+                 (p/catch (fn [error-resp-ex] {:result :client-error
+                                               :value 10
+                                               :resp (ex-data error-resp-ex)}))
+                 (p/timeout 1000 {:result :timeout
+                                  :value 100})
+                 (p/then #(update % :value inc))
+                 (deref))))
+      (is (p/done? req))
+      (is (p/resolved? req))
+      (is (not (p/rejected? req)))
+      (is (not (p/cancelled? req)))
+      (server/shutdown server)))
+  (testing "after timeout"
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          _ (server/start server nil)
+          req (p/promise (server/send-request server "req" {:body "foo"}))]
+      (is (= {:result :timeout
+              :value 101}
+             (-> req
+                 (p/then (fn [resp] {:result :client-success
+                                     :value 1
+                                     :resp resp}))
+                 (p/catch (fn [error-resp-ex] {:result :client-error
+                                               :value 10
+                                               :resp (ex-data error-resp-ex)}))
+                 (p/timeout 300 {:result :timeout
+                                 :value 100})
+                 (p/then #(update % :value inc))
+                 (deref))))
+      (is (not (p/done? req)))
+      (is (not (p/resolved? req)))
+      (is (not (p/rejected? req)))
+      (is (not (p/cancelled? req)))
+      (server/shutdown server)))
+  (testing "after client error"
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          _ (server/start server nil)
+          req (p/promise (server/send-request server "req" {:body "foo"}))
+          client-rcvd-msg (h/assert-take output-ch)]
+      (async/put! input-ch
+                  (-> (lsp.responses/response (:id client-rcvd-msg))
+                      (lsp.responses/error {:code 1234
+                                            :message "Something bad"
+                                            :data {:body "foo"}})))
+      (is (= {:result :client-error
+              :value 11
+              :resp {:jsonrpc "2.0",
+                     :id 1,
+                     :error {:code 1234,
+                             :message "Something bad",
+                             :data {:body "foo"}}}}
+             (-> req
+                 (p/then (fn [resp] {:result :client-success
+                                     :value 1
+                                     :resp resp}))
+                 (p/catch (fn [error-resp-ex] {:result :client-error
+                                               :value 10
+                                               :resp (ex-data error-resp-ex)}))
+                 (p/timeout 1000 {:result :timeout
+                                  :value 100})
+                 (p/then #(update % :value inc))
+                 (deref))))
+      (is (p/done? req))
+      (is (not (p/resolved? req)))
+      (is (p/rejected? req))
+      (is (not (p/cancelled? req)))
+      (server/shutdown server)))
+  (testing "after cancellation"
+    (let [input-ch (async/chan 3)
+          output-ch (async/chan 3)
+          server (server/chan-server {:output-ch output-ch
+                                      :input-ch input-ch})
+          _ (server/start server nil)
+          req (p/promise (server/send-request server "req" {:body "foo"}))]
+      (h/assert-take output-ch)
+      (p/cancel! req)
+      (is (p/done? req))
+      (is (p/cancelled? req))
+      (is (= {:jsonrpc "2.0", :method "$/cancelRequest", :params {:id 1}}
+             (h/assert-take output-ch)))
       (server/shutdown server))))
 
 (def fixed-clock